requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.7
pillow==10.1.0